feat: T1-1 记忆系统重构(FTS5+文件式) + T2-1 Chat页面拆分
**T1-1：记忆系统重构（向量→文件+FTS5）**
- long-term.ts: 用 SQLite FTS5 虚拟表替代向量 cosine 搜索，删除 embeddingFn/reindexEmbeddings()/vectorSearch()
- 引入 6 分类记忆体系：user/operational/governance/skill/correction/reference
- memory_remember 工具新增 category/topic 参数，写 data/.memory/{category}/{topic}.md 文件
- 新增 CMASTER.md 协议：启动时加载 data/CMASTER.md + data/.memory/MEMORY.md 注入 system prompt
- agent.ts: 懒加载全局指令(_loadGlobalInstructions)，缓存至 session 生命周期结束
- index.ts: 移除 embedding 相关初始化代码（embeddingFn/embeddingProvider）
- types.ts: 删除 embeddingProvider/vectorDb/collectionName 字段
- config/default.yaml: 删除 embeddingModel/embeddingProvider 配置项
- server.ts: 将 /api/memories/reindex 改为 /api/memories/migrate（文件式迁移工具）
- 新增 docs/CMASTER.md.example 作为全局指令模板

**T2-1：Chat 页面拆分（727行→6个组件）**
- web/src/components/chat/code-block.tsx: CodeBlock + MermaidHighlighter + MarkdownText
- web/src/components/chat/message-actions.tsx: CopyButton/ReloadButton/FeedbackButton/SuggestionButtons/InlineCopyButton
- web/src/components/chat/interrupt-card.tsx: HitL 高危操作确认卡
- web/src/components/chat/messages.tsx: CustomAssistantMessage/CustomUserMessage/CustomUserEditComposer
- web/src/components/chat/thread-utils.tsx: ThreadHydrator/PromptAutoSender
- web/src/app/chat/page.tsx: 精简至 ~100行，只保留 ChatSession/ChatContent/ChatPage 入口
